service_backend/requirements/base.txt
# Base
pytz==2019.3
python-slugify==4.0.0
Pillow==7.0.0
psycopg2==2.8.4 --no-binary psycopg2

# Django
django==2.2.12

# Environment
django-environ==0.4.5

# JWT
pyjwt==1.7.1

# Django REST Framework
djangorestframework==3.11.0

# Django Rest Password Reset
django-rest-passwordreset==1.1.0

django-filter==2.2.0

# Passwords security
argon2-cffi==19.2.0

# Static files
# whitenoise==5.0.1

# Django CORS headers
django-cors-headers==3.2.1

# Requests
requests==2.23.0

#django-letsencrypt
django-letsencrypt==3.0.1

#friendly errors
drf-friendly-errors==0.13

#Req for core service
numpy==1.19.0
opencv-python==4.5.1.48
fingerprint_enhancer

